Massive refactor and combined tokenization
main.py got overfilled with functions.  Added a model manager to handle model-building, config manager to handle pathing and tensorflow/cuda management, and dataset manager to handle building and inspecting the dataset.

Tokenization is now in one step, at least for the time being.  Will look for better ways to manage vocabulary in the future.